+++ src/message/style/index.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-relative-packages
-import '../../_common/style/mobile/components/message/_index.less';
+import '../../_common/style/mobile/components/message/v2/_index.less';