+++ src/switch/type.ts
@@ -5,19 +5,15 @@
  * */
 
 import { TNode } from '../common';
+import { MouseEvent } from 'react';
 
-export interface TdSwitchProps {
+export interface TdSwitchProps<T = SwitchValue> {
   /**
-   * 自定义颜色，[打开时的颜色，关闭时的颜色]。组件默认颜色为 ['#0052d9', 'rgba(0, 0, 0, .26']。示例：[blue, gray]
-   */
-  colors?: string[];
-  /**
-   * 开关内容，[打开时的值，关闭时的值]。默认为 [true, false]。示例：[1, 0]
+   * 用于自定义开关的值，[打开时的值，关闭时的值]。默认为 [true, false]。示例：[1, 0]、['open', 'close']
    */
   customValue?: Array<SwitchValue>;
   /**
-   * 是否禁用组件
-   * @default false
+   * 是否禁用组件。优先级：Switch.disabled > Form.disabled
    */
   disabled?: boolean;
   /**
@@ -38,15 +34,15 @@ export interface TdSwitchProps {
   /**
    * 开关值
    */
-  value?: SwitchValue;
+  value?: T;
   /**
    * 开关值，非受控属性
    */
-  defaultValue?: SwitchValue;
+  defaultValue?: T;
   /**
    * 数据发生变化时触发
    */
-  onChange?: (value: SwitchValue) => void;
+  onChange?: (value: T, context: { e: MouseEvent<HTMLDivElement> }) => void;
 }
 
 export type SwitchValue = string | number | boolean;