+++ src/textarea/_example/autosize.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Textarea } from 'tdesign-mobile-react';
-
-export default function Autosize() {
-  return <Textarea autosize placeholder="请输入文字" />;
-}