+++ src/textarea/_example/label.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Textarea } from 'tdesign-mobile-react';
+
+export default function Base() {
+  return <Textarea className="textarea-example" label="标签文字" placeholder="请输入文字" />;
+}